refactor instructorCourseAdminInstances
